added comments.
There was a bunch of work that has been scrapped on doing a more intelligent
disabling of bluetooth, something that covers the corner case where someone
connects a device during the period of the bluetooth scan.  This turns out to
be nearly impossible, because android will not give you a list of useful info,
you have to do an active scan and try to connect to things - there appears to be no
bluetooth state information accessible by the app.