removed the hashmap from the loginsessionmanager and all associated code.
AsyncPostSender is ill-conceived and is being entirely rewritten.  It has been refactored for
clarity and with many a TODO, but should not have any functional change.  A new class,
SimpleAsync, has been created, future code should use this class instead.  AsyncPostSender
will be removed once old code is purged.

ForgotPassword was broken to begin with, it is being rewritten to use the new SimpleAsync
class.